TINYMCE-14507: refactor
