+++ modules/oxide-components/src/main/ts/bespoke/tinymceai/tag/Tag.stories.tsx
@@ -1,6 +1,6 @@
 import { Fun } from '@ephox/katamari';
 import type { Meta, StoryObj } from '@storybook/react-vite';
-import { UniverseProvider, type UniverseResources } from 'oxide-components/main';
+import { createTooltipTrigger, UniverseProvider, type UniverseResources } from 'oxide-components/main';
 
 import { Tag } from './Tag';
 
@@ -21,6 +21,7 @@ const resources: UniverseResources = {
     <path fill="#222F3E" fill-rule="evenodd" d="M11.723 5.62 9.356 8l2.367 2.38a.95.95 0 0 1-1.343 1.343L8 9.356l-2.38 2.367a.95.95 0 0 1-1.343-1.343L6.644 8 4.277 5.62A.95.95 0 0 1 5.62 4.277L8 6.644l2.38-2.367a.95.95 0 0 1 1.343 1.343Z"/>
   </svg>
 `),
+  currentTooltipTrigger: createTooltipTrigger()
 };
 
 export const ClosableTag: Story = {
