+++ modules/oxide-components/src/main/ts/contexts/UniverseContext/UniverseTypes.tsx
@@ -1,3 +1,16 @@
+
+export interface TooltipTriggerRef {
+  readonly get: () => string | null;
+  readonly set: (value: string | null) => void;
+  readonly subscribe: (callback: () => void) => () => void;
+}
+export interface TooltipTriggerRef {
+  readonly get: () => string | null;
+  readonly set: (value: string | null) => void;
+  readonly subscribe: (callback: () => void) => () => void;
+}
+
 export interface UniverseResources {
   readonly getIcon: (name: string) => string;
+  readonly currentTooltipTrigger: TooltipTriggerRef;
 };
