+++ modules/oxide-components/src/main/ts/main.ts
@@ -17,11 +17,11 @@ import * as Menu from './components/menu/Menu';
 import * as MenuRenderer from './components/menu/MenuRenderer';
 import * as SegmentedControl from './components/segmentedcontrol/SegmentedControl';
 import { ToolbarInputForm } from './components/toolbarInputForm/ToolbarInputForm';
-import { createTooltipTrigger, type TooltipTriggerRef } from './components/tooltip/internals/Context';
 import * as Tooltip from './components/tooltip/Tooltip';
 import { useUniverse } from './contexts/UniverseContext/Universe';
+import { createTooltipTrigger } from './contexts/UniverseContext/UniverseContext';
 import { UniverseProvider } from './contexts/UniverseContext/UniverseProvider';
-import type { UniverseResources } from './contexts/UniverseContext/UniverseTypes';
+import type { UniverseResources, TooltipTriggerRef } from './contexts/UniverseContext/UniverseTypes';
 import * as KeyboardNavigationTypes from './keynav/keyboard/NavigationTypes';
 import * as KeyboardNavigationHooks from './keynav/KeyboardNavigationHooks';
 import * as Bem from './utils/Bem';
