+++ modules/oxide-components/src/test/ts/browser/bespoke/tinymceai/Tag.spec.tsx
@@ -1,5 +1,5 @@
 import { Fun } from '@ephox/katamari';
-import { Bem, Tag, UniverseProvider, type UniverseResources } from 'oxide-components/main';
+import { Bem, createTooltipTrigger, Tag, UniverseProvider, type UniverseResources } from 'oxide-components/main';
 import { describe, expect, it, vi } from 'vitest';
 import { userEvent } from 'vitest/browser';
 import { render } from 'vitest-browser-react';
@@ -9,6 +9,7 @@ const resources: UniverseResources = {
     <path fill="#222F3E" fill-rule="evenodd" d="M11.723 5.62 9.356 8l2.367 2.38a.95.95 0 0 1-1.343 1.343L8 9.356l-2.38 2.367a.95.95 0 0 1-1.343-1.343L6.644 8 4.277 5.62A.95.95 0 0 1 5.62 4.277L8 6.644l2.38-2.367a.95.95 0 0 1 1.343 1.343Z"/>
   </svg>
 `),
+  currentTooltipTrigger: createTooltipTrigger()
 };
 
 describe('browser.bespoke.tinymceai.Tag', () => {
