+++ modules/oxide-components/src/test/ts/browser/components/ExpandableBox.spec.tsx
@@ -1,6 +1,6 @@
 import { Fun } from '@ephox/katamari';
 import { ExpandableBox, type ExpandableBoxProps } from 'oxide-components/components/expandablebox/ExpandableBox';
-import { UniverseProvider } from 'oxide-components/main';
+import { createTooltipTrigger, UniverseProvider } from 'oxide-components/main';
 import * as Bem from 'oxide-components/utils/Bem';
 import { useState } from 'react';
 import { describe, expect, it, vi } from 'vitest';
@@ -11,6 +11,7 @@ describe('browser.components.ExpandableBoxTest', () => {
   const getIcon = vi.fn((icon: string) => `<svg id="${icon}"></svg>`);
   const mockUniverse = {
     getIcon,
+    currentTooltipTrigger: createTooltipTrigger()
   };
 
   const wrapper = ({ children }: { children: React.ReactNode }) => {
