+++ modules/oxide-components/src/test/ts/browser/components/MenuDivider.spec.tsx
@@ -1,5 +1,6 @@
 import { Fun } from '@ephox/katamari';
 import * as Menu from 'oxide-components/components/menu/Menu';
+import { createTooltipTrigger } from 'oxide-components/contexts/UniverseContext/UniverseContext';
 import { UniverseProvider } from 'oxide-components/contexts/UniverseContext/UniverseProvider';
 import * as Bem from 'oxide-components/utils/Bem';
 import { describe, expect, it } from 'vitest';
@@ -7,6 +8,7 @@ import { render } from 'vitest-browser-react';
 
 const mockUniverse = {
   getIcon: Fun.constant(''),
+  currentTooltipTrigger: createTooltipTrigger()
 };
 
 describe('browser.MenuDividerTest', () => {
