+++ modules/oxide-components/src/test/ts/browser/components/MenuItem.spec.tsx
@@ -1,6 +1,7 @@
 /* eslint-disable max-len */
 import type { CommonMenuItemInstanceApi } from 'oxide-components/components/menu/internals/Types';
 import * as Menu from 'oxide-components/components/menu/Menu';
+import { createTooltipTrigger } from 'oxide-components/contexts/UniverseContext/UniverseContext';
 import { UniverseProvider } from 'oxide-components/contexts/UniverseContext/UniverseProvider';
 import * as Bem from 'oxide-components/utils/Bem';
 import { describe, expect, it, vi } from 'vitest';
@@ -35,6 +36,7 @@ const iconResolver = (icon: string): string => {
 
 const mockUniverse = {
   getIcon: iconResolver,
+  currentTooltipTrigger: createTooltipTrigger()
 };
 
 const waitForElementText = async (getByText: (text: string) => Locator, text: string) => {
